fourier transform impl AND float to double

+++ src/AudioOut.hpp
@@ -1,5 +1,5 @@
 // AudioOut.hpp
 #ifndef AUDIO_H
 #define AUDIO_H
-void initAudio(float* wave_table);
+void initAudio(double* wave_table);
 #endif
